bkp_17th_Oct_2025/agents/requirements.txt
azure-ai-projects==1.0.0b11
azure-core==1.32.0
azure-identity==1.20.0
python-dotenv==1.0.1
jsonref==1.1.0